a bit more cleaning
